refactor(auto-matching): rename layout and update import paths
