+++ src/app/(private)/lobby/(auto-leave)/auto-leave-handler.tsx
