+++ src/app/(private)/lobby/(auto-leave)/auto-matching/client-component.tsx
@@ -9,7 +9,7 @@ import { Tiny } from '~/app/_fonts';
 import { WaitingText } from '~/components/ui';
 import { routes } from '~/constants/routes';
 import { useMainGameSocket } from '~/socket';
-import { UserCard } from '../_components/user-card';
+import { UserCard } from '../../_components/user-card';
 import type { MatchingMode } from './page';
 
 interface Props {
